feat: implement proactive token refresh scheduler with graceful shutdown handling
