+++ src/app.js
@@ -26,15 +26,10 @@ if(process.env.NODE_ENV !== 'production'){
   app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(specs));
 }
 
-const port = process.env.NODE_ENV === 'production'
-             ? process.env.BACKEND_PROD_PORT
-             : process.env.BACKEND_DEV_PORT;
-
 const deviceRouter = require('./routes/device.route')
 const serversRouter = require('./routes/servers.route')
 const streamRouter = require('./routes/stream.route')
 const healthRouter = require('./routes/health.route')
-const { refreshIfExpiringSoon } = require('./services/device.service');
 
 // Accept all sources of connection requests. This is to accommodate requests coming from both rs.local:3000 and from the rshake device ip address
 app.use(cors())
@@ -45,41 +40,6 @@ app.use('/device', deviceRouter)
 app.use('/servers', serversRouter)
 app.use('/stream', streamRouter)
 app.use('/health', healthRouter)
-
-// Proactive token refresh scheduler
-let refreshTimer = null;
-const clearRefreshTimer = () => {
-  if (refreshTimer) {
-    clearTimeout(refreshTimer);
-    refreshTimer = null;
-  }
-};
-
-const scheduleRefresh = async () => {
-  try {
-    await refreshIfExpiringSoon();
-  } catch (error) {
-    console.log(`Proactive token refresh failed: ${error.message || error}`);
-  } finally {
-    const interval = Number(process.env.REFRESH_CHECK_INTERVAL_MS || 15 * 60 * 1000);
-    refreshTimer = setTimeout(scheduleRefresh, interval);
-  }
-};
-
-if (process.env.NODE_ENV !== 'test') {
-  scheduleRefresh();
-
-  const handleShutdownSignal = (signal) => {
-    console.log(`Received ${signal}, clearing proactive token refresh timer.`);
-    clearRefreshTimer();
-    // Re-emit so default shutdown behaviour still happens.
-    process.kill(process.pid, signal);
-  };
-
-  process.once('SIGINT', () => handleShutdownSignal('SIGINT'));
-  process.once('SIGTERM', () => handleShutdownSignal('SIGTERM'));
-  process.once('exit', clearRefreshTimer);
-}
 
 /* Error handler middleware */
 app.use((err, req, res, next) => {
